requirements.txt
google-genai==0.3.0
google-cloud-storage==2.18.2
requests==2.32.3
PyYAML==6.0.2
tqdm==4.66.4
Pillow==10.4.0
python-slugify==8.0.4
